撮影・保存処理をViewModelに移行
