+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="theme">お題: %sの表情</string>
     <string name="take_photo">Take Photo</string>
     <string name="camera_permission_message">カメラへのアクセスを許可してください</string>
+    <string name="failed_save_image">画像の保存に失敗しました</string>
 </resources>
